requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
datasets==3.5.1
dill==0.3.8
distro==1.9.0
effdet==0.4.1
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
fugashi==1.4.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
iopath==0.1.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
layoutparser==0.3.4
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
omegaconf==2.3.0
openai==1.74.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pluggy==1.5.0
portalocker==3.1.1
propcache==0.3.1
pyarrow==20.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
pymupdf4llm==0.0.21
pynndescent==0.5.13
pyparsing==3.2.3
pypdfium2==4.30.1
pysbd==0.3.4
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
unidic-lite==1.0.8
urllib3==2.4.0
vila @ git+https://github.com/allenai/vila.git@f482972b1d8e1377afc866494f488de81b1d71e5
xxhash==3.5.0
yarl==1.20.0
